openssh: refactor ssh config
Seapath ssh config in bbappend has the same file name than openssh ssh
config. Thus the original configuration of openssh is overwritten anyway
even if DISTRO_FEATURE is checked.

- Use a different name to avoid for ssh config in the bbappend
- Remove seapath-ro overrides as this distro is not used
- Modify ssh configuration only if seapath-security or seapath-readonly
are set.

Issue: #6860

+++ recipes-connectivity/openssh/openssh_%.bbappend
@@ -4,27 +4,27 @@
 FILESEXTRAPATHS_prepend := "${THISDIR}/files:"
 
 SRC_URI += " \
-    file://sshd_config \
-    file://sshd_config_flash \
-    file://sshd_config_readonly \
+    file://sshd_config_seapath \
+    file://sshd_config_seapath_flash \
+    file://sshd_config_seapath_readonly \
 "
 
 do_install_append() {
-    install -d ${D}${sysconfdir}/ssh
-    install -m 0644 ${WORKDIR}/sshd_config \
-       ${D}${sysconfdir}/ssh
+    if ${@bb.utils.contains('DISTRO_FEATURES','seapath-readonly','true','false',d)}; then
+        install -d ${D}${sysconfdir}/ssh
+        install -m 0644 ${WORKDIR}/sshd_config_seapath_readonly \
+           ${D}${sysconfdir}/ssh/sshd_config
+    elif ${@bb.utils.contains('DISTRO_FEATURES','seapath-security','true','false',d)}; then
+        install -d ${D}${sysconfdir}/ssh
+        install -m 0644 ${WORKDIR}/sshd_config_seapath \
+           ${D}${sysconfdir}/ssh/sshd_config
+    fi
 }
 
 do_install_append_seapath-flash() {
     install -d ${D}${sysconfdir}/ssh
-    install -m 0644 ${WORKDIR}/sshd_config_flash \
+    install -m 0644 ${WORKDIR}/sshd_config_seapath_flash \
         ${D}${sysconfdir}/ssh/sshd_config
-}
-
-do_install_append_seapath-ro() {
-    install -d ${D}${sysconfdir}/ssh
-    install -m 0644 ${WORKDIR}/sshd_config_readonly \
-       ${D}${sysconfdir}/ssh
 }
 
 FILES_${PN} += " \
